fix: change location of burger factory
